pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eodhd-mcp-server"
version = "1.0.0"
description = "EODHD API MCP Server for accessing financial data through Model Context Protocol"
authors = [{name = "EODHD MCP Server", email = "info@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
eodhd-mcp-server = "eodhd_mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/example/eodhd-mcp-server"
Repository = "https://github.com/example/eodhd-mcp-server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"